Moved all of the component styling to one file. #7
