+++ client/story-card.js
@@ -3,24 +3,12 @@ import { default as OverFlowEllipsis } from 'react-overflow-ellipsis'
 import Card, { CardActions, CardContent, CardMedia } from 'material-ui/Card'
 import Button from 'material-ui/Button'
 import Typography from 'material-ui/Typography'
+import styles from './styles/styles'
 
-const styles = {
-  card: {
-    maxWidth: 345,
-    maxHeight: 450,
-    marginTop: 20
-  },
-  captions: {
-    marginTop: 10
-  },
-  media: {
-    height: 200
-  }
-}
 export default function StoryCard({ story, seeMoreInfo }) {
   return (
     <div className="col-1-4">
-      <Card className="content" style={styles.card}>
+      <Card className="content" style={styles.storyCard}>
         <CardMedia
           style={styles.media}
           image={story.image}
