+++ client/story-details.js
@@ -3,47 +3,7 @@ import Card, { CardContent } from 'material-ui/Card'
 import Paper from 'material-ui/Paper'
 import Typography from 'material-ui/Typography'
 import AudioPlayer from './audio-player'
-
-const styles = {
-  card: {
-    margin: 50,
-    padding: 30
-  },
-  header: {
-    marginLeft: 32,
-    marginRight: 55,
-    padding: 20
-  },
-  title: {
-    marginLeft: 10,
-    marginRight: 30,
-    marginTop: 10,
-    fontSize: 50,
-    fontWeight: 700
-  },
-  author: {
-    marginLeft: 10,
-    marginRight: 30,
-    marginTop: 10,
-    fontSize: 15,
-    fontWeight: 500
-  },
-  content: {
-    marginTop: 30,
-    marginLeft: 15,
-    marginRight: 40,
-    padding: 20
-  },
-  contentText: {
-    marginLeft: 10,
-    marginRight: 20
-  },
-  image: {
-    maxWidth: 1000,
-    maxHeight: 1000,
-    marginLeft: 15
-  }
-}
+import styles from './styles/styles'
 
 export default function StoryDetails({ story }) {
   return (
