+++ client/story-form.js
@@ -4,42 +4,7 @@ import TextField from 'material-ui/TextField'
 import Typography from 'material-ui/Typography'
 import Dropzone from 'react-dropzone'
 import Button from 'material-ui/Button'
-
-const styles = {
-  usersForm: {
-    marginLeft: 100,
-    marginRight: 100,
-    marginTop: 50,
-    paddingLeft: 100,
-    paddingRight: 100,
-    paddingTop: 40,
-    textAlign: 'center'
-  },
-  textField: {
-    marginTop: 30
-  },
-  submit: {
-    marginTop: 30,
-    marginBottom: 30
-  },
-  dropzone: {
-    display: 'inline-block',
-    width: 1000,
-    marginTop: 30,
-    textAlign: 'center',
-    fontFamily: 'Roboto'
-  },
-  image: {
-    width: 300,
-    marginRight: 20,
-    border: '1px dashed black'
-  },
-  audio: {
-    width: 300,
-    marginLeft: 20,
-    border: '1px dashed black'
-  }
-}
+import styles from './styles/styles'
 
 function onDrop(files) {
   console.log(files)
@@ -47,7 +12,7 @@ function onDrop(files) {
 
 export default function StoryForm({ handleStorySubmit }) {
   return (
-    <Paper style={styles.usersForm}>
+    <Paper style={styles.storyForm}>
       <Typography type="headline" component="h1">Upload a Story!</Typography>
       <form onSubmit={handleStorySubmit}>
         <div>
@@ -55,7 +20,7 @@ export default function StoryForm({ handleStorySubmit }) {
             type="text"
             label="Title"
             name="title"
-            style={styles.textField}
+            style={styles.storyInputField}
             fullWidth={true}
           />
         </div>
@@ -66,21 +31,21 @@ export default function StoryForm({ handleStorySubmit }) {
             label="Content"
             multiline
             rows="8"
-            style={styles.textField}
+            style={styles.storyInputField}
             margin="normal"
             fullWidth={true}
           />
         </div>
         <div className="dropzone" style={styles.dropzone}>
-          <Dropzone accept="image/jpeg, image/png" onDrop={onDrop} name="image" className="col-1-2" style={styles.image}>
+          <Dropzone accept="image/jpeg, image/png" onDrop={onDrop} name="image" className="col-1-2" style={styles.imageDrop}>
             <p>Upload an Image</p>
           </Dropzone>
-          <Dropzone accept="audio/*" onDrop={onDrop} name="audio" className="col-1-2" style={styles.audio}>
+          <Dropzone accept="audio/*" onDrop={onDrop} name="audio" className="col-1-2" style={styles.audioDrop}>
             <p>Upload an Audio File</p>
           </Dropzone>
         </div>
         <div>
-          <Button raised color="primary" type="submit" style={styles.submit}>Submit</Button>
+          <Button raised color="primary" type="submit" style={styles.submitStory}>Submit</Button>
         </div>
       </form>
     </Paper>
