+++ client/styles/styles.js
@@ -1,0 +1,182 @@
+const styles = {
+  root: {
+    marginTop: 10,
+    width: '100%'
+  },
+  flex: {
+    flex: 1
+  },
+  menuButton: {
+    marginLeft: -12,
+    marginRight: 20
+  },
+  audioPlayer: {
+    position: 'fixed',
+    bottom: 0,
+    width: '100%',
+    backgroundColor: '#4054b1',
+    marginRight: 300,
+    marginTop: 30,
+    padding: 30,
+    borderRadius: 5
+  },
+  audioTitle: {
+    color: 'white',
+    textDecoration: 'underline',
+    textAlign: 'center',
+    marginBottom: 20
+  },
+  audioButtons: {
+    position: 'relative',
+    top: 3,
+    marginLeft: 50,
+    color: 'white'
+  },
+  play: {
+    minWidth: 50
+  },
+  pause: {
+    minWidth: 50
+  },
+  stop: {
+    marginLeft: 20
+  },
+  durationBar: {
+    position: 'relative',
+    bottom: 10,
+    width: 700,
+    marginLeft: 50
+  },
+  volumeIcon: {
+    minWidth: 50,
+    position: 'relative',
+    top: 3,
+    color: 'white',
+    marginLeft: 50
+  },
+  volumeBar: {
+    position: 'relative',
+    bottom: 10,
+    width: 150,
+    marginLeft: 20
+  },
+  storyCard: {
+    maxWidth: 345,
+    maxHeight: 450,
+    marginTop: 20
+  },
+  captions: {
+    marginTop: 10
+  },
+  media: {
+    height: 200
+  },
+  card: {
+    margin: 50,
+    padding: 30
+  },
+  header: {
+    marginLeft: 32,
+    marginRight: 55,
+    padding: 20
+  },
+  title: {
+    marginLeft: 10,
+    marginRight: 30,
+    marginTop: 10,
+    fontSize: 50,
+    fontWeight: 700
+  },
+  author: {
+    marginLeft: 10,
+    marginRight: 30,
+    marginTop: 10,
+    fontSize: 15,
+    fontWeight: 500
+  },
+  content: {
+    marginTop: 30,
+    marginLeft: 15,
+    marginRight: 40,
+    padding: 20
+  },
+  contentText: {
+    marginLeft: 10,
+    marginRight: 20
+  },
+  image: {
+    maxWidth: 1000,
+    maxHeight: 1000,
+    marginLeft: 15
+  },
+  storyForm: {
+    marginLeft: 100,
+    marginRight: 100,
+    marginTop: 50,
+    paddingLeft: 100,
+    paddingRight: 100,
+    paddingTop: 40,
+    textAlign: 'center'
+  },
+  storyInputField: {
+    marginTop: 30
+  },
+  submitStory: {
+    marginTop: 30,
+    marginBottom: 30
+  },
+  dropzone: {
+    display: 'inline-block',
+    width: 1000,
+    marginTop: 30,
+    textAlign: 'center',
+    fontFamily: 'Roboto'
+  },
+  imageDrop: {
+    width: 300,
+    marginRight: 20,
+    border: '1px dashed black'
+  },
+  audioDrop: {
+    width: 300,
+    marginLeft: 20,
+    border: '1px dashed black'
+  },
+  paper: {
+    marginTop: 50,
+    padding: 30
+  },
+  searchField: {
+    width: 950,
+    marginTop: 30
+  },
+  searchBar: {
+    textAlign: 'center'
+  },
+  icon: {
+    marginLeft: 30,
+    marginRight: 10
+  },
+  postButton: {
+    marginLeft: 45,
+    marginRight: 40
+  },
+  usersForm: {
+    marginLeft: 100,
+    marginRight: 100,
+    marginTop: 50,
+    paddingLeft: 100,
+    paddingRight: 100,
+    paddingTop: 40,
+    textAlign: 'center'
+  },
+  userInputField: {
+    marginTop: 30
+  },
+  submitUser: {
+    marginTop: 30,
+    marginBottom: 30
+  }
+}
+
+export default styles
